requirements.txt
PyQt6>=6.6.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
aiohttp>=3.9.0
